Create README.md
## Project Overview: A Serverless Automated ETL Data Pipeline

This project is a powerful, end-to-end serverless data pipeline designed to solve a critical public health challenge in Cameroon: ensuring children receive timely vaccinations.
 Inspired by personal experiences, this solution provides a robust and automated way to manage patient data and send proactive reminders to healthcare providers, all while running on a highly scalable and cost-effective cloud infrastructure.

The pipeline is triggered by the upload of patient data (CSV files) to an Amazon S3 data lake. 
An AWS Glue Workflow orchestrates a Glue Job that performs an ETL process, transforming the raw data to identify children due for vaccination. 
This job then leverages Amazon SNS to send targeted email notifications to a Public Relations Officer (PRO). 

The data is made available for on-demand analysis and visualization through Amazon Athena and Amazon QuickSight, empowering healthcare teams with data-driven insights. The entire system is secured by IAM and monitored with CloudWatch, guaranteeing reliability and security without the need for manual intervention.